requirements.txt
# Core Application & ASGI Server Stack
fastapi==0.115.12
pydantic==2.11.4
uvicorn[standard]==0.34.2
gunicorn==23.0.0

# PDF Ingestion and Parsing
pdfminer.six[image]==20250506 # User confirmed this version
# System libs for Pillow (libjpeg-dev zlib1g-dev libtiff5-dev) will be handled on VM

# External API Interaction & Authentication
httpx==0.28.1
google-auth==2.38.0

# Caching Mechanisms
cachecontrol[filecache]==0.14.3

# Essential Utilities
python-dotenv==1.1.0

# Optional: API Security
# python-jose[cryptography]==3.4.0 # Keep if you anticipate needing it